chore(folder-structure): moves /app folder into src folder
