+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
-import "@/app/ui/global.css";
-import PageContainer from "@/app/ui/PageContainer";
-import Footer from "@/app/ui/Footer";
+import "@/src/app/ui/global.css";
+import PageContainer from "@/src/app/ui/PageContainer";
+import Footer from "@/src/app/ui/Footer";
 import { Navbar } from "./ui/Navbar";
 import Script from "next/script";
