+++ src/app/ui/blog/components/Details.tsx
