+++ src/app/ui/blog/components/FigureImage.tsx
