+++ src/app/ui/blog/components/Table.tsx
